+++ WorkoutApp/Repository/OrderRepository.cs
@@ -51,7 +51,7 @@ namespace WorkoutApp.Repository
         }
 
         /// <inheritdoc/>
-        public async Task<Order> GetByIdAsync(long id)
+        public async Task<Order> GetByIdAsync(int id)
         {
             const string query = "SELECT * FROM [Order] WHERE ID = @ID";
             var parameters = new List<SqlParameter>
@@ -116,7 +116,7 @@ namespace WorkoutApp.Repository
         }
 
         /// <inheritdoc/>
-        public async Task<bool> DeleteAsync(long id)
+        public async Task<bool> DeleteAsync(int id)
         {
             const string query = "UPDATE [Order] SET IsActive = 0 WHERE ID = @ID";
             var parameters = new List<SqlParameter>
